requirements.txt
django == 5.1.1
pip == 24.1.2
djangorestframework == 3.15.2